Move prefix extractor out of filter policy mod
